# openapi.yaml
openapi: 3.0.3
info:
  title: University Services API
  version: 1.0.0
  description: 'The central API platform for university operations.

    Covers Authentication, Students, Programs & Courses, Faculty, Library, AI
    Assistant, API Key Management, and UAE Government Reporting domains.


    ## Authentication

    Most endpoints require a Bearer JWT token obtained via `POST /auth/login`.

    Government reporting endpoints (`/gov/*`) require an `X-API-Key` header instead.


    Get a JWT token: `POST /auth/login` with your email and password.

    Use JWT: Add header `Authorization: Bearer <token>` to all requests.

    Use API Key: Add header `X-API-Key: uak_<your_key>` — service role only.


    ## Test Credentials (local / seeded environment)

    | Role | Email | Password |

    |------|-------|----------|

    | Admin | admin@university.ac.ae | Admin@1234 |

    | Student | student1@university.ac.ae | Student@1234 |

    | Faculty | faculty1@university.ac.ae | Faculty@1234 |

    | Librarian | librarian@university.ac.ae | Librarian@1234 |


    ## API-Led Connectivity Architecture

    Follows MuleSoft API-led connectivity:

    - **System Layer**: Raw data access (student records, course catalog, book inventory)

    - **Process Layer**: Business logic (enrollment validation, borrowing rules, capacity
    checks)

    - **Experience Layer**: Consumer-tailored responses (student portal, faculty dashboard)

...


    ## AI Library Assistant

...
    OpenAI.

...

    '
...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained via OAuth2 authorization flow
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
...
  schemas:
...
    ApiKeyCreate:
      type: object
      required:
      - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
          example: MOEHE Integration
        role:
          type: string
          enum: [service, admin, faculty, student, librarian]
          default: service
        expires_in_days:
          type: integer
          minimum: 1
          maximum: 3650
          default: 365
          description: Days until expiry. Omit for no expiry.
    ApiKeyUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        is_active:
          type: boolean
        expires_at:
          type: string
          format: date-time
    ApiKeyResponse:
      type: object
      description: Safe representation — never includes key hash.
      properties:
        id:
          type: string
        name:
          type: string
          example: MOEHE Integration
        key_prefix:
          type: string
          example: uak_3f8a2c1b
          description: First 12 chars of the key — for identification only.
        role:
          type: string
          example: service
        is_active:
          type: boolean
        expires_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ApiKeyCreateResponse:
      type: object
      description: Returned only on create and rotate. Raw key shown once — store immediately.
      properties:
        key:
          type: string
          example: uak_3f8a2c1b9e4d...
          description: Raw API key — store this securely, shown only once.
        api_key:
          $ref: '#/components/schemas/ApiKeyResponse'
    ApiKeyListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/ApiKeyResponse'
        total:
          type: integer

    # ── Sprint 9 — UAE Gov schemas ──────────────────────────────────────────
...
    ErrorResponse:
      type: object
      required:
      - error
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: The request body contains invalid fields
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  issue:
                    type: string
              example:
              - field: email
                issue: Must be a valid university email address
...
  responses:
...
    Unauthorized:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Authenticated but insufficient permissions for this action
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
security:
- BearerAuth: []
paths:
...
  /api-keys:
    post:
      tags:
      - API Keys
      summary: Create API key
      description: Generates a new API key. The raw key is returned exactly once — store it securely. Admin only.
      operationId: createApiKey
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApiKeyCreate'
            example:
              name: MOEHE Integration
              role: service
              expires_in_days: 365
      responses:
        '201':
          description: API key created. Raw key shown once — copy immediately.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyCreateResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    get:
      tags:
      - API Keys
      summary: List API keys
      description: Returns all API keys. Never includes key hashes — only prefix and metadata. Admin only.
      operationId: listApiKeys
      responses:
        '200':
          description: List of API keys
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api-keys/{key_id}:
    patch:
      tags:
      - API Keys
      summary: Update API key
      description: Update name, active status, or expiry. Cannot change role or regenerate hash. Admin only.
      operationId: updateApiKey
      parameters:
      - name: key_id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApiKeyUpdate'
      responses:
        '200':
          description: Updated API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags:
      - API Keys
      summary: Revoke API key
      description: Soft-deletes by setting is_active=false. Row kept for audit trail. Admin only.
      operationId: revokeApiKey
      parameters:
      - name: key_id
        in: path
        required: true
        schema:
          type: string
      responses:
        '204':
          description: Key revoked
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api-keys/{key_id}/rotate:
    post:
      tags:
      - API Keys
      summary: Rotate API key
      description: Generates a new key and replaces the stored hash atomically. Old key stops working immediately. Returns new raw key once.
      operationId: rotateApiKey
      parameters:
      - name: key_id
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: New raw key — copy immediately, shown only once.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyCreateResponse'
        '400':
          description: Cannot rotate a revoked key
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── UAE GOVERNMENT ──────────────────────────────────────────────────────────
...